Extract border specific colors Tailwind plugin to its own file
