+++ tailwind.config.js
@@ -1,7 +1,10 @@
-const flattenColorPalette = require('tailwindcss/lib/util/flattenColorPalette').default;
+const borderSpecificColors = require('./lib/tailwind_plugins/border_specific_colors.js');
 
 module.exports = {
   purge: false,
+  plugins: [
+    borderSpecificColors
+  ],
   theme: {
     variants: {
       borderWidth: ['responsive', 'hover']
@@ -187,26 +190,5 @@ module.exports = {
         DEFAULT: theme('colors.primary')
       })
     }
-  },
-  plugins: [
-    ({
-      addUtilities, _e, theme, variants
-    }) => {
-      // Border specific colours
-      // https://github.com/tailwindlabs/tailwindcss/pull/560#issuecomment-670045304
-      const colors = flattenColorPalette(theme('borderColor'));
-      delete colors.default;
-
-      const colorMap = Object.keys(colors)
-        .map((color) => ({
-          [`.border-t-${color}`]: { borderTopColor: colors[color] },
-          [`.border-r-${color}`]: { borderRightColor: colors[color] },
-          [`.border-b-${color}`]: { borderBottomColor: colors[color] },
-          [`.border-l-${color}`]: { borderLeftColor: colors[color] }
-        }));
-      const utilities = Object.assign({}, ...colorMap);
-
-      addUtilities(utilities, variants('borderColor'));
-    }
-  ]
+  }
 };
